backend/foodgram/requirements.txt
asgiref==3.6.0
Django==3.2.16
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
